% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{looks_rtf}
\alias{looks_rtf}
\title{Quick test whether a file is RTF}
\usage{
looks_rtf(con, n = 1000)
}
\arguments{
\item{con}{A connection object or string of file name}

\item{n}{Integer that specifies the length of contents to be tested.
If smaller than 10, forced to 10.}
}
\value{
Logical.
}
\description{
Conducts a quick test whether a file looks RTF.
}
