% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated-functions.R
\name{striprtf}
\alias{rtf2text}
\alias{striprtf}
\title{Renamed Functions}
\usage{
striprtf(file, verbose = FALSE, ...)

rtf2text(text, verbose = FALSE)
}
\description{
From ver 0.3.1, the functions are renamed as follows:
\itemize{
\item{\code{striprtf} --> \code{\link{read_rtf}}}
\item{\code{rtf2text} --> \code{\link{strip_rtf}}}
}
}

