% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/striprtf.R
\name{striprtf}
\alias{rtf2text}
\alias{striprtf}
\title{Extract Text from RTF (Rich Text Format) File}
\usage{
striprtf(file)

rtf2text(text)
}
\arguments{
\item{file}{Path to an RTF file. Must be character of length 1.}

\item{text}{Character of length 1.  Expected to be contents of an RTF file.}
}
\value{
character of extracted text
}
\description{
Extract Text from RTF (Rich Text Format) File
}
\references{
https://gist.github.com/gilsondev/7c1d2d753ddb522e7bc22511cfb08676
}

