% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{looks_rtf}
\alias{looks_rtf}
\title{Test if a file looks like an RTF}
\usage{
looks_rtf(con, n = 1000)
}
\arguments{
\item{con}{A connection object or string of file name}

\item{n}{Integer that specifies the length of contents to be tested.
If smaller than 10, forced to 10.}
}
\value{
Logical.
}
\description{
Validate if a file looks like an RTF.
The test should be seen as a minimal requirement;
If failed, the file is highly likely that the file is invalid,
while passed, there is still possibility that the file does not follw the rule of RTF files.
}
