% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated-functions.R
\name{striprtf-deprecated}
\alias{striprtf-deprecated}
\alias{striprtf}
\alias{rtf2text}
\title{Renamed Functions}
\usage{
striprtf(file, verbose = FALSE, ...)

rtf2text(text, verbose = FALSE)
}
\arguments{
\item{file}{Path to an RTF file. Must be character of length 1.}

\item{verbose}{Logical. If TRUE, progress report is printed on console.
While it can be informative when parsing a large file,
this option itself makes the process slow.}

\item{...}{Addional arguments passed to \code{\link{readLines}}}

\item{text}{Character of length 1.  Expected to be contents of an RTF file.}
}
\value{
Character vector of extracted text
}
\description{
From ver 0.3.1, the functions are renamed as follows:
\itemize{
\item{\code{striprtf} --> \code{\link{read_rtf}}}
\item{\code{rtf2text} --> \code{\link{strip_rtf}}}
}
}
