% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/striprtf.R
\name{striprtf}
\alias{rtf2text}
\alias{striprtf}
\title{Extract Text from RTF (Rich Text Format) File}
\usage{
striprtf(file, verbose = FALSE, ...)

rtf2text(text, verbose = FALSE)
}
\arguments{
\item{file}{Path to an RTF file. Must be character of length 1.}

\item{verbose}{Logical. If TRUE, progress report is printed on console.
While it can be informative when parsing a large file,
this option itself makes the process slow.}

\item{...}{Addional arguments passed to \code{\link{readLines}}}

\item{text}{Character of length 1.  Expected to be contents of an RTF file.}
}
\value{
Character vector of extracted text
}
\description{
Extract Text from RTF (Rich Text Format) File
}
\examples{
striprtf(system.file("extdata/king.rtf", package = "striprtf"))
}
\references{
\itemize{
 \item{https://gist.github.com/gilsondev/7c1d2d753ddb522e7bc22511cfb08676}
 \item{http://www.unicode.org/Public/MAPPINGS/VENDORS/MICSFT/WINDOWS/}
}
}

