% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/striprtf.R
\name{striprtf}
\alias{rtf2text}
\alias{striprtf}
\title{Extract Text from RTF (Rich Text Format) File}
\usage{
striprtf(file, verbose = FALSE, ...)

rtf2text(text, verbose = FALSE)
}
\arguments{
\item{file}{Path to an RTF file. Must be character of length 1.}

\item{verbose}{Logical. If TRUE, progress report is shown.}

\item{...}{Addional arguments passed to \code{\link{readLines}}}

\item{text}{Character of length 1.  Expected to be contents of an RTF file.}
}
\value{
Character vector of extracted text
}
\description{
Extract Text from RTF (Rich Text Format) File
}

